+++ Magic/CHANGELOG.md
@@ -37,6 +37,7 @@
  - Fix ChangeContext not being able to set the source entity to nothing (affects Mob, Monster, etc spells)
  - Fix a whole pile of memory leaks
  - Improved support for persistent mobs (in 1.14 and up)
+ - Add "land" trigger and "fall_distance" attribute
 
 # 8.1.2